PLTable v1.0.2
Package name changed from prettytable to pltable. PLTable can be used as a drop-in replacement for PTable, provided that the module import name is changed accordingly. The opposite is not true because PLTable adds additional capabilities which are not present at PTable, while retaining the existing with bug fixes when necessary.